PCheck/requirements.txt
asgiref==3.9.1
channels==4.1.0
colorama==0.4.6
crispy-bootstrap5==2025.6
daphne==4.1.2
Django==5.2
django-allauth==0.62.0
django-crispy-forms==2.4
django-widget-tweaks==1.5.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
qrcode==8.2
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
